Surveillance
I am interested in observing how people move through a space,
especially when given delayed visual feedback on their interactions
with the projection and within the space itself. This program stores
the previous frame pixel information and compares it to the current
frame, showing the differences between the two with color manipulation
on a 5 second time delay; this allows for delayed visual feedback of
the space and therefore a form of delayed gratification. As humans, we
rely primarily on our visual perception; we seek immediate visual
information or stimulation. In this way, the webcam acts as an antenna;
it receives visual information through surveillance of the space and
transmits that information to be manipulated.